feat: implement comprehensive PDF statement generation with structured layout, enhanced font handling, and improved header/footer management for better user experience
