+++ internal/pdf/builder.go
@@ -1,999 +1,13 @@
 package pdf
 
 import (
-	"bytes"
-	_ "embed"
-	"fmt"
-	"image"
-	_ "image/jpeg"
-	_ "image/png"
-	"strings"
 	"time"
 
-	"github.com/jung-kurt/gofpdf/v2"
-	"github.com/shopspring/decimal"
-
+	"github.com/stellar/stellar-disbursement-platform-backend/internal/pdf/statement"
 	"github.com/stellar/stellar-disbursement-platform-backend/internal/services"
-	"github.com/stellar/stellar-disbursement-platform-backend/internal/utils"
 )
-
-//go:embed assets/fonts/Inter_24pt-Regular.ttf
-var interRegularFont []byte
-
-//go:embed assets/fonts/Inter_24pt-Bold.ttf
-var interBoldFont []byte
-
-//go:embed assets/fonts/Inter_24pt-Medium.ttf
-var interMediumFont []byte
-
-//go:embed assets/fonts/Inter_24pt-SemiBold.ttf
-var interSemiBoldFont []byte
-
-//go:embed assets/fonts/GoogleSansCode-Regular.ttf
-var googleSansCodeRegularFont []byte
-
-// useMockTransactions enables mock statement data (multiple currencies + transactions) for PDF testing. Set to false for production.
-const useMockTransactions = false
-
-// Page dimensions and margins
-const (
-	mmPerPage    = 210.0
-	pageHeight   = 297.0
-	marginLR     = 15.0
-	marginTop    = 15.0
-	marginBottom = 25.0
-)
-
-// Font sizes
-const (
-	bodyFontSize          = 8.0
-	tableHeaderSize       = 8.0
-	sectionTitleSize      = 12.0
-	txCellFontSize        = 8.0
-	txSmallFontSize       = 6.7
-	organizationNameFontSize = 9.2
-	dateRangeFontSize     = 10.0
-	titleFontSize         = 23.4
-)
-
-// Line heights
-const (
-	summaryHeaderLineHeight  = 4.0
-	headerLeftColLineHeight  = 6.0
-	headerRightColLineHeight = 5.0
-	footerLineHeight         = 4.0
-	titleSectionLine1Height  = 6.0
-	titleSectionLine2Height  = 9.0
-	titleSectionLine3Height  = 6.0
-	dateLineHeight           = 4.0
-	counterpartyLineHeight   = 4.0
-	txIDLineHeight           = 4.0
-	txHeaderLineHeight       = 4.0
-	amountLineHeight         = 4.0
-	currencyLineHeight       = 4.0
-)
-
-// Table row heights
-const (
-	summaryHeaderRowHeight  = 15.5
-	summaryDataRowHeight    = 14
-	txHeaderRowHeight       = 15.5
-	txDataRowHeight         = 16
-)
-
-// Table cell spacing
-const (
-	cellPaddingX    = 2.115
-	counterpartyGap = 1.0
-)
-
-// Section margins
-const (
-	sectionBottomMargin = 12.0
-)
-
-// Header section spacing
-const (
-	headerBottomMargin              = 5.0
-	miniHeaderBottomMargin          = 5.0
-	headerSeparatorBottomMargin     = 6.0
-	miniHeaderSeparatorBottomMargin = 3.0
-	headerSeparatorLineWidth        = 0.26
-	headerLogoToOrgNameGap          = 3.0
-	logoOffsetX                     = 1.0
-	walletAddressLabelGap           = 0.3
-)
-
-// Title section spacing
-const (
-	titleSectionBottomMargin = 5.0
-)
-
-// Footer section spacing
-const (
-	footerMarginTop           = 1.5
-	footerContentGap          = 3.5
-	footerDisclaimerToPageGap = 2.0
-)
-
-// Other constants
-const (
-	walletIDLabel             = "Wallet ID"
-	maxCounterpartyTextLength = 35
-)
-
-// tableWidth is full content width between left and right margins (both = marginLR).
-const tableWidth = mmPerPage - 2*marginLR
-
-// summaryColWidths: Wallet Address | Beginning Balance | Total Credits | Total Debits | Ending Balance (sum = tableWidth)
-var summaryColWidths = []float64{44, 34, 34, 34, 34}
-
-// txColWidths: Date | Transaction ID | Counterparty | Debits | Credits | Balance (sum = tableWidth)
-var txColWidths = []float64{16, 52, 49, 21, 21, 21}
-
-// Text colors (RGB)
-var headerAndTotalsColor = []int{54, 65, 83}   // #364153
-var defaultCellColor      = []int{74, 85, 101}  // #4A5565
-var noteColor = []int{106, 114, 130} // #6A7282
-var activeColor    = []int{20, 71, 230}  // #1447E6
-var highlightColor    = []int{16, 24, 40}   // #101828
-var sectionTitleColor    = []int{30, 41, 57}   // #1E2939
-
-// Border and background colors (RGB)
-var headerBorderColor = []int{203, 207, 215}   // #CBCFD7
-var defaultBorderColor = []int{240, 241, 243}     // #F0F1F3
-var totalsRowBgColor = []int{249, 250, 251}   // #F9FAFB
-
 
 // BuildPDF generates a multi-page PDF from a StatementResult and returns the bytes.
 func BuildPDF(result *services.StatementResult, fromDate, toDate time.Time, organizationName string, organizationLogo []byte) ([]byte, error) {
-	pdf := gofpdf.New("P", "mm", "A4", "")
-
-	pdf.AddUTF8FontFromBytes("Inter", "", interRegularFont)
-	pdf.AddUTF8FontFromBytes("Inter", "B", interBoldFont)
-	pdf.AddUTF8FontFromBytes("Inter", "M", interMediumFont)
-	pdf.AddUTF8FontFromBytes("Inter", "emi", interSemiBoldFont) // SemiBold: register as "emi" (no "S") to avoid gofpdf strikethrough
-	pdf.AddUTF8FontFromBytes("GoogleSansCode", "", googleSansCodeRegularFont)
-
-	pdf.SetMargins(marginLR, marginTop, marginLR)
-	pdf.SetAutoPageBreak(true, marginBottom)
-	pdf.SetFooterFunc(func() {
-		// Position at top of footer area; add margin, then top border).
-		pdf.SetY(pageHeight - marginBottom)
-		pdf.Ln(footerMarginTop)
-		pdf.SetDrawColor(defaultBorderColor[0], defaultBorderColor[1], defaultBorderColor[2])
-		pdf.SetLineWidth(headerSeparatorLineWidth)
-		y := pdf.GetY()
-		pdf.Line(marginLR, y, mmPerPage-marginLR, y)
-		pdf.SetY(y + 1.0)
-		pdf.Ln(footerContentGap)
-		pdf.SetLineWidth(0.25)
-		pdf.SetDrawColor(0, 0, 0)
-
-		// Disclaimer and page count fit within marginBottom (same as marginTop).
-		// Disclaimer label and text using absolute positioning
-		yDisclaimerLine := pdf.GetY()
-		disclaimerLabel := "Disclaimer:"
-		disclaimerText := " This report is generated from SDP records. Blockchain confirmations reflect public ledger data."
-		// Draw label in bold
-		pdf.SetFont("Inter", "B", bodyFontSize)
-		pdf.SetTextColor(noteColor[0], noteColor[1], noteColor[2])
-		labelWidth := pdf.GetStringWidth(disclaimerLabel)
-		xLabelStart := pdf.GetX()
-		pdf.SetXY(xLabelStart, yDisclaimerLine)
-		pdf.CellFormat(labelWidth, footerLineHeight, disclaimerLabel, "", 0, "L", false, 0, "")
-		// Calculate text position
-		xTextStart := xLabelStart + labelWidth
-		// Draw text in regular font
-		pdf.SetFont("Inter", "", bodyFontSize)
-		pdf.SetXY(xTextStart, yDisclaimerLine)
-		pdf.CellFormat(0, footerLineHeight, disclaimerText, "", 1, "L", false, 0, "")
-		pdf.Ln(footerDisclaimerToPageGap)
-		pdf.CellFormat(0, footerLineHeight, fmt.Sprintf("Page %d of %d", pdf.PageNo(), pdf.PageCount()), "", 0, "R", false, 0, "")
-		pdf.SetTextColor(0, 0, 0)
-	})
-	pdf.AddPage()
-
-	if useMockTransactions {
-		result = getMockStatementResult()
-	}
-	walletAccount := result.Summary.Account
-	drawHeader(pdf, organizationName, organizationLogo, fromDate, toDate)
-	drawTitleSection(pdf, walletAccount)
-	drawHeaderSeparatorLine(pdf, headerSeparatorBottomMargin)
-
-	pdf.SetFont("Inter", "B", sectionTitleSize)
-	pdf.SetTextColor(sectionTitleColor[0], sectionTitleColor[1], sectionTitleColor[2])
-	pdf.CellFormat(0, 8, "Account Summary", "", 1, "L", false, 0, "")
-	pdf.SetTextColor(0, 0, 0)
-	pdf.Ln(1)
-
-	pdf.SetFont("Inter", "emi", tableHeaderSize)
-	pdf.SetTextColor(headerAndTotalsColor[0], headerAndTotalsColor[1], headerAndTotalsColor[2])
-	pdf.SetDrawColor(headerBorderColor[0], headerBorderColor[1], headerBorderColor[2])
-	pdf.SetLineWidth(0.53)
-	ySummaryHeaderStart := pdf.GetY()
-	xPos := pdf.GetX()
-	xSummaryLeft := xPos
-	summaryHeaders := []struct {
-		text  string
-		align string
-	}{
-		{"Currency", "L"},
-		{"Beginning Balance", "R"},
-		{"Total Credits", "R"},
-		{"Total Debits", "R"},
-		{"Ending Balance", "R"},
-	}
-	for i, h := range summaryHeaders {
-		w := summaryColWidths[i]
-		pdf.SetXY(xPos, ySummaryHeaderStart)
-		pdf.CellFormat(w, summaryHeaderRowHeight, "", "B", 0, "L", false, 0, "")
-		textW := w - 2*cellPaddingX
-		if i == 0 {
-			pdf.SetXY(xPos+cellPaddingX, ySummaryHeaderStart)
-			pdf.CellFormat(textW, summaryHeaderRowHeight, h.text, "", 0, h.align, false, 0, "")
-		} else {
-			lines := strings.Split(breakHeaderWords(h.text), "\n")
-			lineHeight := summaryHeaderLineHeight
-			blockHeight := lineHeight * float64(len(lines))
-			lineY := ySummaryHeaderStart + (summaryHeaderRowHeight-blockHeight)/2
-			for _, line := range lines {
-				pdf.SetXY(xPos+cellPaddingX, lineY)
-				pdf.CellFormat(textW, summaryHeaderLineHeight, line, "", 0, "R", false, 0, "")
-				lineY += lineHeight
-			}
-		}
-		xPos += w
-	}
-	pdf.SetXY(xSummaryLeft, ySummaryHeaderStart+summaryHeaderRowHeight)
-	pdf.SetLineWidth(0.25)
-	pdf.SetDrawColor(0, 0, 0)
-	pdf.SetTextColor(0, 0, 0)
-	pdf.SetFont("Inter", "", bodyFontSize)
-
-	pdf.SetDrawColor(defaultBorderColor[0], defaultBorderColor[1], defaultBorderColor[2])
-	pdf.SetLineWidth(0.26)
-	for _, asset := range result.Summary.Assets {
-		xPos = xSummaryLeft
-		for _, w := range summaryColWidths {
-			pdf.SetXY(xPos, pdf.GetY())
-			pdf.CellFormat(w, summaryDataRowHeight, "", "B", 0, "L", false, 0, "")
-			xPos += w
-		}
-		ySummaryData := pdf.GetY()
-		textW0 := summaryColWidths[0] - 2*cellPaddingX
-		pdf.SetXY(xSummaryLeft+cellPaddingX, ySummaryData)
-		pdf.SetFont("Inter", "B", bodyFontSize)
-		pdf.SetTextColor(highlightColor[0], highlightColor[1], highlightColor[2])
-		pdf.CellFormat(textW0, summaryDataRowHeight, asset.Code, "", 0, "L", false, 0, "")
-		xPos = xSummaryLeft + summaryColWidths[0]
-		for i, w := range summaryColWidths[1:] {
-			textW := w - 2*cellPaddingX
-			var text string
-			switch i {
-			case 0:
-				text = utils.FormatAmountTo2Decimals(asset.BeginningBalance)
-			case 1:
-				text = utils.FormatAmountTo2Decimals(asset.TotalCredits)
-			case 2:
-				text = utils.FormatAmountTo2Decimals(asset.TotalDebits)
-			case 3:
-				text = utils.FormatAmountTo2Decimals(asset.EndingBalance)
-			}
-			if i < 3 {
-				pdf.SetFont("Inter", "", bodyFontSize)
-				pdf.SetTextColor(highlightColor[0], highlightColor[1], highlightColor[2])
-			} else {
-				pdf.SetFont("Inter", "emi", bodyFontSize)
-				pdf.SetTextColor(activeColor[0], activeColor[1], activeColor[2])
-			}
-			pdf.SetXY(xPos+cellPaddingX, ySummaryData)
-			pdf.CellFormat(textW, summaryDataRowHeight, text, "", 0, "R", false, 0, "")
-			xPos += w
-		}
-		pdf.SetFont("Inter", "", bodyFontSize)
-		pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-		pdf.SetXY(xSummaryLeft, ySummaryData+summaryDataRowHeight)
-	}
-	pdf.SetTextColor(0, 0, 0)
-	pdf.SetLineWidth(0.25)
-	pdf.SetDrawColor(0, 0, 0)
-	pdf.Ln(sectionBottomMargin)
-
-	pageBottom := pageHeight - marginBottom
-	for _, asset := range result.Summary.Assets {
-		if len(asset.Transactions) == 0 {
-			continue
-		}
-		pdf.SetFont("Inter", "B", sectionTitleSize)
-		pdf.SetTextColor(sectionTitleColor[0], sectionTitleColor[1], sectionTitleColor[2])
-		pdf.CellFormat(0, 8, "Transactions - "+asset.Code, "", 1, "L", false, 0, "")
-		pdf.SetTextColor(0, 0, 0)
-		pdf.Ln(1)
-
-		drawTxTableHeader(pdf)
-		runningBalance, err := decimal.NewFromString(asset.BeginningBalance)
-		if err != nil {
-			runningBalance = decimal.Zero
-		}
-		transactions := asset.Transactions
-		for i := range transactions {
-			if pdf.GetY()+txDataRowHeight > pageBottom {
-				pdf.AddPage()
-				if pdf.PageNo() > 1 {
-					drawMiniHeader(pdf, organizationName, organizationLogo, fromDate, toDate, walletAccount)
-				}
-				drawTxTableHeader(pdf)
-			}
-			runningBalance = drawTxRow(pdf, &transactions[i], asset.Code, runningBalance)
-		}
-		if pdf.GetY()+txDataRowHeight > pageBottom {
-			pdf.AddPage()
-			if pdf.PageNo() > 1 {
-				drawMiniHeader(pdf, organizationName, organizationLogo, fromDate, toDate, walletAccount)
-			}
-			drawTxTableHeader(pdf)
-		}
-		drawTotalsRowForAsset(pdf, &asset)
-		pdf.Ln(sectionBottomMargin)
-	}
-
-	var buf bytes.Buffer
-	if err := pdf.Output(&buf); err != nil {
-		return nil, fmt.Errorf("writing PDF: %w", err)
-	}
-	return buf.Bytes(), nil
-}
-
-func drawTxTableHeader(pdf *gofpdf.Fpdf) {
-	pdf.SetFont("Inter", "emi", tableHeaderSize)
-	pdf.SetTextColor(headerAndTotalsColor[0], headerAndTotalsColor[1], headerAndTotalsColor[2])
-	pdf.SetDrawColor(headerBorderColor[0], headerBorderColor[1], headerBorderColor[2])
-	pdf.SetLineWidth(0.4)
-	yStart := pdf.GetY()
-	xPos := pdf.GetX()
-	txHeaders := []struct {
-		text  string
-		align string
-	}{
-		{"Date", "L"},
-		{"Transaction ID", "L"},
-		{"Counterparty", "L"},
-		{"Debits", "R"},
-		{"Credits", "R"},
-		{"Balance", "R"},
-	}
-	for i, h := range txHeaders {
-		w := txColWidths[i]
-		pdf.SetXY(xPos, yStart)
-		pdf.CellFormat(w, txHeaderRowHeight, "", "B", 0, "L", false, 0, "")
-		textW := w - 2*cellPaddingX
-		if h.align == "L" {
-			pdf.SetXY(xPos+cellPaddingX, yStart)
-			pdf.CellFormat(textW, txHeaderRowHeight, h.text, "", 0, h.align, false, 0, "")
-		} else {
-			lines := strings.Split(breakHeaderWords(h.text), "\n")
-			blockHeight := txHeaderLineHeight * float64(len(lines))
-			lineY := yStart + (txHeaderRowHeight-blockHeight)/2
-			for _, line := range lines {
-				pdf.SetXY(xPos+cellPaddingX, lineY)
-				pdf.CellFormat(textW, txHeaderLineHeight, line, "", 0, "R", false, 0, "")
-				lineY += txHeaderLineHeight
-			}
-		}
-		xPos += w
-	}
-	pdf.SetY(yStart + txHeaderRowHeight)
-	pdf.SetLineWidth(0.25)
-	pdf.SetDrawColor(0, 0, 0)
-	pdf.SetTextColor(0, 0, 0)
-	pdf.SetFont("Inter", "", bodyFontSize)
-}
-
-// drawTxRow draws one transaction row and returns the new running balance after this transaction.
-func drawTxRow(pdf *gofpdf.Fpdf, tx *services.StatementTransaction, assetCode string, runningBalance decimal.Decimal) decimal.Decimal {
-	var dateLine1, dateLine2 string
-	dateStr := tx.UpdatedAt
-	if dateStr == "" {
-		dateStr = tx.CreatedAt
-	}
-	parsedTime, err := time.Parse(time.RFC3339, dateStr)
-	if err == nil {
-		utcTime := parsedTime.UTC()
-		dateLine1 = utcTime.Format("2006-01-02")
-		dateLine2 = utcTime.Format("15:04 UTC")
-	} else {
-		if len(dateStr) >= 10 {
-			dateLine1 = dateStr[:10]
-		} else {
-			dateLine1 = dateStr
-		}
-		dateLine2 = ""
-	}
-	opID := tx.ID
-
-	// Counterparty: Recipient/sender wallet information
-	walletAddr := utils.TruncateString(tx.CounterpartyAddress, 5)
-	name := tx.CounterpartyName
-	hasName := name != ""
-	var line2Label, line2Value string
-	if hasName {
-		if tx.Type == "credit" {
-			// For credit rows, don't show "Sender" label
-			line2Value = name
-		} else {
-			// For debit rows, show "Recipient • name"
-			line2Label = "Recipient"
-			line2Value = name
-		}
-		if line2Label != "" && len(line2Label+" • "+line2Value) > maxCounterpartyTextLength {
-			combined := line2Label + " • " + line2Value
-			combined = utils.TruncateToMaxLength(combined, maxCounterpartyTextLength)
-			// Try to preserve the structure, but truncate if needed
-			parts := strings.Split(combined, " • ")
-			if len(parts) >= 2 {
-				line2Label = parts[0]
-				line2Value = parts[1]
-			} else {
-				line2Value = combined
-				line2Label = ""
-			}
-		} else if line2Value != "" && len(line2Value) > maxCounterpartyTextLength {
-			line2Value = utils.TruncateToMaxLength(line2Value, maxCounterpartyTextLength)
-		}
-	}
-
-	var debitsAmount, creditsAmount string
-	amount, _ := decimal.NewFromString(tx.Amount)
-	amountStr := utils.FormatAmountTo2Decimals(tx.Amount)
-	if tx.Type == "debit" {
-		debitsAmount = amountStr
-		runningBalance = runningBalance.Sub(amount)
-	} else {
-		creditsAmount = amountStr
-		runningBalance = runningBalance.Add(amount)
-	}
-	balanceAmountStr := utils.FormatDecimal(runningBalance)
-
-	// Draw cells
-	pdf.SetDrawColor(defaultBorderColor[0], defaultBorderColor[1], defaultBorderColor[2])
-	pdf.SetLineWidth(0.26)
-	pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-	xStart := pdf.GetX()
-	yStart := pdf.GetY()
-
-	// Date
-	pdf.SetFont("Inter", "", txSmallFontSize)
-	pdf.SetXY(xStart, yStart)
-	pdf.CellFormat(txColWidths[0], txDataRowHeight, "", "B", 0, "L", false, 0, "")
-	dateCellWidth := txColWidths[0] - 2*cellPaddingX
-	dateBlockHeight := dateLineHeight * 2
-	dateY := yStart + (txDataRowHeight-dateBlockHeight)/2
-	// First line: date
-	pdf.SetXY(xStart+cellPaddingX, dateY)
-	pdf.CellFormat(dateCellWidth, dateLineHeight, dateLine1, "", 0, "L", false, 0, "")
-	// Second line: time and UTC
-	if dateLine2 != "" {
-		pdf.SetXY(xStart+cellPaddingX, dateY+dateLineHeight)
-		pdf.CellFormat(dateCellWidth, dateLineHeight, dateLine2, "", 0, "L", false, 0, "")
-	}
-
-	xID := xStart + txColWidths[0]
-	pdf.SetXY(xID, yStart)
-	pdf.CellFormat(txColWidths[1], txDataRowHeight, "", "B", 0, "L", false, 0, "")
-	cellWidth := txColWidths[1] - 2*cellPaddingX
-	txURL := fmt.Sprintf("https://stellar.expert/explorer/testnet/tx/%s", tx.ID)
-	pdf.SetFont("GoogleSansCode", "U", txSmallFontSize)
-	pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-	idLines := pdf.SplitText(opID, cellWidth)
-	idBlockHeight := float64(len(idLines)) * txIDLineHeight
-	var extBlockHeight float64
-	if tx.ExternalPaymentID != "" {
-		extBlockHeight = txIDLineHeight
-	}
-	totalBlockHeight := idBlockHeight + extBlockHeight
-	blockY := yStart + (txDataRowHeight-totalBlockHeight)/2
-	pdf.SetXY(xID+cellPaddingX, blockY)
-	pdf.MultiCell(cellWidth, txIDLineHeight, opID, "", "L", false)
-	pdf.LinkString(xID+cellPaddingX, blockY, cellWidth, idBlockHeight, txURL)
-	if tx.ExternalPaymentID != "" {
-		pdf.SetFont("Inter", "", txSmallFontSize)
-		pdf.SetXY(xID+cellPaddingX, blockY+idBlockHeight)
-		pdf.CellFormat(cellWidth, txIDLineHeight, tx.ExternalPaymentID, "", 0, "L", false, 0, "")
-	}
-	pdf.SetY(yStart)
-	pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-
-	// Counterparty
-	const bulletChar = "•"
-	xCP := xStart + txColWidths[0] + txColWidths[1]
-	cpW := txColWidths[2] - 2*cellPaddingX
-	pdf.SetXY(xCP, yStart)
-	pdf.CellFormat(txColWidths[2], txDataRowHeight, "", "B", 0, "L", false, 0, "")
-	xTextStart := xCP + cellPaddingX
-
-	// Calculate vertical centering for 2-line content
-	counterpartyBlockHeight := counterpartyLineHeight * 2
-	counterpartyY := yStart + (txDataRowHeight-counterpartyBlockHeight)/2
-
-	pdf.SetFont("Inter", "", txSmallFontSize)
-
-	if tx.Type == "credit" {
-		// Credit (sender): Line 1 = "Sender Wallet Address" •, Line 2 = wallet address
-		// First line: Label • bullet
-		pdf.SetTextColor(noteColor[0], noteColor[1], noteColor[2])
-		line1Text := "Sender Wallet Address •"
-		pdf.SetXY(xTextStart, counterpartyY)
-		pdf.CellFormat(cpW, counterpartyLineHeight, line1Text, "", 0, "L", false, 0, "")
-
-		// Second line: Wallet address with link
-		pdf.SetFont("GoogleSansCode", "U", txSmallFontSize)
-		pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-		pdf.SetXY(xTextStart, counterpartyY+counterpartyLineHeight)
-		walletURL := fmt.Sprintf("https://stellar.expert/explorer/testnet/account/%s", tx.CounterpartyAddress)
-		pdf.CellFormat(cpW, counterpartyLineHeight, walletAddr, "", 0, "L", false, 0, walletURL)
-		pdf.LinkString(xTextStart, counterpartyY+counterpartyLineHeight, cpW, counterpartyLineHeight, walletURL)
-	} else {
-		// Debit (recipient)
-		if hasName {
-			// First line: Label
-			line1Label := "Wallet Address •"
-			pdf.SetFont("Inter", "", txSmallFontSize)
-			pdf.SetTextColor(noteColor[0], noteColor[1], noteColor[2])
-			labelWidth := pdf.GetStringWidth(line1Label)
-			xLabelStart := xTextStart
-			pdf.SetXY(xLabelStart, counterpartyY)
-			pdf.CellFormat(labelWidth, counterpartyLineHeight, line1Label, "", 0, "L", false, 0, "")
-			
-			// Calculate wallet address position
-			xWalletStart := xLabelStart + labelWidth + counterpartyGap
-			// Wallet address with link
-			pdf.SetFont("GoogleSansCode", "U", txSmallFontSize)
-			pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-			walletURL := fmt.Sprintf("https://stellar.expert/explorer/testnet/account/%s", tx.CounterpartyAddress)
-			walletWidth := cpW - (xWalletStart - xTextStart)
-			pdf.SetXY(xWalletStart, counterpartyY)
-			pdf.CellFormat(walletWidth, counterpartyLineHeight, walletAddr, "", 0, "L", false, 0, walletURL)
-			pdf.LinkString(xWalletStart, counterpartyY, walletWidth, counterpartyLineHeight, walletURL)
-
-			// Second line: Recipient name
-			pdf.SetFont("Inter", "", txSmallFontSize)
-			pdf.SetTextColor(noteColor[0], noteColor[1], noteColor[2])
-			line2Text := "Recipient • " + line2Value
-			pdf.SetXY(xTextStart, counterpartyY+counterpartyLineHeight)
-			pdf.CellFormat(cpW, counterpartyLineHeight, line2Text, "", 0, "L", false, 0, "")
-		} else {
-			// First line: Label
-			singleLineY := yStart + (txDataRowHeight-counterpartyLineHeight)/2
-			line1Label := "Wallet Address • "
-			pdf.SetFont("Inter", "", txSmallFontSize)
-			pdf.SetTextColor(noteColor[0], noteColor[1], noteColor[2])
-			labelWidth := pdf.GetStringWidth(line1Label)
-			xLabelStart := xTextStart
-			pdf.SetXY(xLabelStart, singleLineY)
-			pdf.CellFormat(labelWidth, counterpartyLineHeight, line1Label, "", 0, "L", false, 0, "")
-			
-			// Calculate wallet address position
-			xWalletStart := xLabelStart + labelWidth
-			// Wallet address with link
-			pdf.SetFont("GoogleSansCode", "U", txSmallFontSize)
-			pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-			walletURL := fmt.Sprintf("https://stellar.expert/explorer/testnet/account/%s", tx.CounterpartyAddress)
-			walletWidth := cpW - (xWalletStart - xTextStart)
-			pdf.SetXY(xWalletStart, singleLineY)
-			pdf.CellFormat(walletWidth, counterpartyLineHeight, walletAddr, "", 0, "L", false, 0, walletURL)
-			pdf.LinkString(xWalletStart, singleLineY, walletWidth, counterpartyLineHeight, walletURL)
-		}
-	}
-	// Debits
-	xDebits := xStart + txColWidths[0] + txColWidths[1] + txColWidths[2]
-	if debitsAmount != "" {
-		drawAmountWithCurrency(pdf, amountCellArgs{xDebits, yStart, txColWidths[3], txDataRowHeight, debitsAmount, assetCode, "B", false, amountCellOpts{amountColor: highlightColor}}, cellPaddingX)
-	} else {
-		pdf.SetXY(xDebits, yStart)
-		pdf.CellFormat(txColWidths[3], txDataRowHeight, "", "B", 0, "R", false, 0, "")
-	}
-	// Credits
-	xCredits := xDebits + txColWidths[3]
-	if creditsAmount != "" {
-		drawAmountWithCurrency(pdf, amountCellArgs{xCredits, yStart, txColWidths[4], txDataRowHeight, creditsAmount, assetCode, "B", false, amountCellOpts{amountColor: highlightColor}}, cellPaddingX)
-	} else {
-		pdf.SetXY(xCredits, yStart)
-		pdf.CellFormat(txColWidths[4], txDataRowHeight, "", "B", 0, "R", false, 0, "")
-	}
-	// Balance
-	xBalance := xCredits + txColWidths[4]
-	drawAmountWithCurrency(pdf, amountCellArgs{xBalance, yStart, txColWidths[5], txDataRowHeight, balanceAmountStr, assetCode, "B", false, amountCellOpts{}}, cellPaddingX)
-
-	pdf.SetXY(xStart, yStart+txDataRowHeight)
-	pdf.SetLineWidth(0.25)
-	pdf.SetDrawColor(0, 0, 0)
-	pdf.SetTextColor(0, 0, 0)
-	pdf.SetFont("Inter", "", bodyFontSize)
-
-	return runningBalance
-}
-
-func drawTotalsRowForAsset(pdf *gofpdf.Fpdf, asset *services.StatementAssetSummary) {
-	pdf.SetFillColor(totalsRowBgColor[0], totalsRowBgColor[1], totalsRowBgColor[2])
-	pdf.SetDrawColor(headerBorderColor[0], headerBorderColor[1], headerBorderColor[2])
-	pdf.SetLineWidth(0.53)
-	xStart := pdf.GetX()
-	yStart := pdf.GetY()
-
-	pdf.SetFont("Inter", "emi", txCellFontSize)
-	pdf.SetTextColor(headerAndTotalsColor[0], headerAndTotalsColor[1], headerAndTotalsColor[2])
-	xRowStart := xStart
-	for i := 0; i < 3; i++ {
-		w := txColWidths[i]
-		pdf.SetXY(xStart, yStart)
-		pdf.CellFormat(w, txDataRowHeight, "", "T", 0, "L", true, 0, "")
-		if i == 2 {
-			textW := w - 2*cellPaddingX
-			pdf.SetXY(xStart+cellPaddingX, yStart)
-			pdf.CellFormat(textW, txDataRowHeight, "Totals:", "", 0, "R", false, 0, "")
-		}
-		xStart += w
-	}
-
-	xDebits := xStart
-	xCredits := xDebits + txColWidths[3]
-	xBalance := xCredits + txColWidths[4]
-
-	drawAmountWithCurrency(pdf, amountCellArgs{xDebits, yStart, txColWidths[3], txDataRowHeight, utils.FormatAmountTo2Decimals(asset.TotalDebits), asset.Code, "T", true, amountCellOpts{forTotals: true}}, cellPaddingX)
-	drawAmountWithCurrency(pdf, amountCellArgs{xCredits, yStart, txColWidths[4], txDataRowHeight, utils.FormatAmountTo2Decimals(asset.TotalCredits), asset.Code, "T", true, amountCellOpts{forTotals: true}}, cellPaddingX)
-	drawAmountWithCurrency(pdf, amountCellArgs{xBalance, yStart, txColWidths[5], txDataRowHeight, utils.FormatAmountTo2Decimals(asset.EndingBalance), asset.Code, "T", true, amountCellOpts{forTotals: true, amountColor: activeColor}}, cellPaddingX)
-
-	pdf.SetXY(xRowStart, yStart+txDataRowHeight)
-	pdf.SetLineWidth(0.25)
-	pdf.SetDrawColor(0, 0, 0)
-	pdf.SetTextColor(0, 0, 0)
-	pdf.SetFont("Inter", "", bodyFontSize)
-}
-
-// amountCellOpts configures drawAmountWithCurrency
-type amountCellOpts struct {
-	forTotals   bool
-	amountColor []int
-}
-
-// amountCellArgs holds arguments for drawAmountWithCurrency.
-type amountCellArgs struct {
-	x, y, w, h       float64
-	amountStr        string
-	currencyCode     string
-	border           string
-	fill             bool
-	opts             amountCellOpts
-}
-
-// drawAmountWithCurrency draws a single cell with fixed height
-func drawAmountWithCurrency(pdf *gofpdf.Fpdf, a amountCellArgs, paddingH float64) {
-	x, y, w, h := a.x, a.y, a.w, a.h
-	opts := a.opts
-	// Draw cell border
-	pdf.SetXY(x, y)
-	pdf.CellFormat(w, h, "", a.border, 0, "R", a.fill, 0, "")
-	textW := w - 2*paddingH
-	textX := x + paddingH
-	// Amount + currency
-	amountCurrencyBlockHeight := amountLineHeight + currencyLineHeight
-	amountY := y + (h-amountCurrencyBlockHeight)/2
-	if opts.forTotals {
-		pdf.SetFont("Inter", "emi", txCellFontSize)
-	} else {
-		pdf.SetFont("Inter", "M", txCellFontSize)
-	}
-	if opts.amountColor != nil {
-		pdf.SetTextColor(opts.amountColor[0], opts.amountColor[1], opts.amountColor[2])
-	}
-	pdf.SetXY(textX, amountY)
-	pdf.CellFormat(textW, amountLineHeight, a.amountStr, "", 0, "R", false, 0, "")
-	if opts.amountColor != nil {
-		pdf.SetTextColor(0, 0, 0)
-	}
-	// Currency
-	pdf.SetTextColor(noteColor[0], noteColor[1], noteColor[2])
-	if opts.forTotals {
-		pdf.SetFont("Inter", "emi", txSmallFontSize)
-	} else {
-		pdf.SetFont("Inter", "", txSmallFontSize)
-	}
-	pdf.SetXY(textX, amountY+amountLineHeight)
-	pdf.CellFormat(textW, currencyLineHeight, a.currencyCode, "", 0, "R", false, 0, "")
-	pdf.SetTextColor(0, 0, 0)
-	pdf.SetFont("Inter", "", bodyFontSize)
-}
-
-// drawHeader draws the header: left column (logo + organization name), right column (generated date, statement period).
-func drawHeader(pdf *gofpdf.Fpdf, organizationName string, organizationLogo []byte, fromDate, toDate time.Time) {
-	xLeft := pdf.GetX()
-	yStart := pdf.GetY()
-	contentWidth := tableWidth
-	halfWidth := contentWidth / 2
-	rightColX := xLeft + halfWidth
-
-	var yLeftBottom float64 = yStart
-
-	// Left column
-	if len(organizationLogo) > 0 {
-		imgName, imgInfo := registerLogoImage(pdf, organizationLogo)
-		if imgName != "" && imgInfo != nil {
-			const logoMaxHeight = 10.0
-			imgW, imgH := imgInfo.Width(), imgInfo.Height()
-			if imgH > logoMaxHeight {
-				imgW = imgW * (logoMaxHeight / imgH)
-				imgH = logoMaxHeight
-			}
-			pdf.ImageOptions(imgName, xLeft+logoOffsetX, yStart, imgW, imgH, false, gofpdf.ImageOptions{}, 0, "")
-			yLeftBottom = yStart + imgH + headerLogoToOrgNameGap
-		}
-	}
-	if organizationName != "" {
-		pdf.SetFont("Inter", "B", organizationNameFontSize)
-		pdf.SetTextColor(highlightColor[0], highlightColor[1], highlightColor[2])
-		pdf.SetXY(xLeft, yLeftBottom)
-		pdf.CellFormat(halfWidth, headerLeftColLineHeight, strings.ToUpper(organizationName), "", 0, "L", false, 0, "")
-		yLeftBottom += headerLeftColLineHeight
-	}
-
-	// Right column
-	pdf.SetFont("Inter", "", bodyFontSize)
-	pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-	genStr := fmt.Sprintf("Generated on %s", time.Now().UTC().Format("2006-01-02 15:04 UTC"))
-	pdf.SetXY(rightColX, yStart)
-	pdf.CellFormat(halfWidth, headerRightColLineHeight, genStr, "", 0, "R", false, 0, "")
-	pdf.SetXY(rightColX, yStart+headerRightColLineHeight)
-	pdf.SetFont("Inter", "emi", bodyFontSize)
-	pdf.SetTextColor(highlightColor[0], highlightColor[1], highlightColor[2])
-	pdf.CellFormat(halfWidth, headerRightColLineHeight, "Statement Period:", "", 0, "R", false, 0, "")
-	periodStr := fmt.Sprintf("%s to %s", fromDate.Format("2006-01-02"), toDate.Format("2006-01-02"))
-	pdf.SetXY(rightColX, yStart+2*headerRightColLineHeight)
-	pdf.SetFont("Inter", "B", dateRangeFontSize)
-	pdf.SetTextColor(highlightColor[0], highlightColor[1], highlightColor[2])
-	pdf.CellFormat(halfWidth, headerRightColLineHeight, periodStr, "", 0, "R", false, 0, "")
-	yRightBottom := yStart + 3*headerRightColLineHeight
-
-	// Advance Y to the bottom of the taller column, then add margin
-	if yRightBottom > yLeftBottom {
-		yLeftBottom = yRightBottom
-	}
-	pdf.SetXY(xLeft, yLeftBottom)
-	pdf.Ln(headerBottomMargin)
-	pdf.SetTextColor(0, 0, 0)
-}
-
-// drawMiniHeader draws a compact header for pages 2+
-func drawMiniHeader(pdf *gofpdf.Fpdf, organizationName string, organizationLogo []byte, fromDate, toDate time.Time, walletAccount string) {
-	xLeft := pdf.GetX()
-	yStart := pdf.GetY()
-	contentWidth := tableWidth
-	halfWidth := contentWidth / 2
-	rightColX := xLeft + halfWidth
-
-	var yLeftBottom float64 = yStart
-	var logoWidth float64
-
-	// Left side, first column
-	if len(organizationLogo) > 0 {
-		imgName, imgInfo := registerLogoImage(pdf, organizationLogo)
-		if imgName != "" && imgInfo != nil {
-			const logoMaxHeight = 10.0
-			imgW, imgH := imgInfo.Width(), imgInfo.Height()
-			if imgH > logoMaxHeight {
-				imgW = imgW * (logoMaxHeight / imgH)
-				imgH = logoMaxHeight
-			}
-			pdf.ImageOptions(imgName, xLeft+logoOffsetX, yStart, imgW, imgH, false, gofpdf.ImageOptions{}, 0, "")
-			logoWidth = imgW + logoOffsetX
-			yLeftBottom = yStart + imgH + headerLogoToOrgNameGap
-		}
-	}
-
-	// Left side, second column
-	col2X := xLeft + logoWidth + headerLogoToOrgNameGap
-	col2Width := halfWidth - logoWidth - headerLogoToOrgNameGap
-	if col2Width > 0 {
-		walletAddr := walletAccount
-		if strings.HasPrefix(walletAddr, "stellar:") {
-			walletAddr = walletAddr[8:]
-		}
-		truncatedAddr := utils.TruncateString(walletAddr, 5)
-
-		// Line 1
-		if organizationName != "" {
-			pdf.SetFont("Inter", "B", organizationNameFontSize)
-			pdf.SetTextColor(highlightColor[0], highlightColor[1], highlightColor[2])
-			pdf.SetXY(col2X, yStart)
-			pdf.CellFormat(col2Width, headerLeftColLineHeight, strings.ToUpper(organizationName), "", 0, "L", false, 0, "")
-		}
-		// Line 2
-		yLine2 := yStart + headerLeftColLineHeight
-		pdf.SetFont("Inter", "", organizationNameFontSize)
-		pdf.SetTextColor(noteColor[0], noteColor[1], noteColor[2])
-		labelText := "Wallet Address: "
-		labelWidth := pdf.GetStringWidth(labelText)
-		pdf.SetXY(col2X, yLine2)
-		pdf.CellFormat(labelWidth, titleSectionLine3Height, labelText, "", 0, "L", false, 0, "")
-		xValueStart := col2X + labelWidth + walletAddressLabelGap
-		pdf.SetFont("GoogleSansCode", "U", organizationNameFontSize)
-		pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-		walletURL := fmt.Sprintf("https://stellar.expert/explorer/testnet/account/%s", walletAddr)
-		truncatedWidth := pdf.GetStringWidth(truncatedAddr)
-		pdf.SetXY(xValueStart, yLine2)
-		pdf.CellFormat(truncatedWidth, titleSectionLine3Height, truncatedAddr, "", 0, "L", false, 0, walletURL)
-		pdf.LinkString(xValueStart, yLine2, truncatedWidth, titleSectionLine3Height, walletURL)
-		textBlockBottom := yLine2 + titleSectionLine3Height
-		if textBlockBottom > yLeftBottom {
-			yLeftBottom = textBlockBottom
-		}
-	}
-
-	// Right column
-	pdf.SetFont("Inter", "", bodyFontSize)
-	pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-	genStr := fmt.Sprintf("Generated on %s", time.Now().UTC().Format("2006-01-02 15:04 UTC"))
-	pdf.SetXY(rightColX, yStart)
-	pdf.CellFormat(halfWidth, headerRightColLineHeight, genStr, "", 0, "R", false, 0, "")
-	pdf.SetXY(rightColX, yStart+headerRightColLineHeight)
-	pdf.SetFont("Inter", "emi", bodyFontSize)
-	pdf.SetTextColor(highlightColor[0], highlightColor[1], highlightColor[2])
-	pdf.CellFormat(halfWidth, headerRightColLineHeight, "Statement Period:", "", 0, "R", false, 0, "")
-	periodStr := fmt.Sprintf("%s to %s", fromDate.Format("2006-01-02"), toDate.Format("2006-01-02"))
-	pdf.SetXY(rightColX, yStart+2*headerRightColLineHeight)
-	pdf.SetFont("Inter", "B", dateRangeFontSize)
-	pdf.SetTextColor(highlightColor[0], highlightColor[1], highlightColor[2])
-	pdf.CellFormat(halfWidth, headerRightColLineHeight, periodStr, "", 0, "R", false, 0, "")
-	yRightBottom := yStart + 3*headerRightColLineHeight
-
-	if yRightBottom > yLeftBottom {
-		yLeftBottom = yRightBottom
-	}
-	pdf.SetXY(xLeft, yLeftBottom)
-	pdf.Ln(miniHeaderBottomMargin)
-	pdf.SetTextColor(0, 0, 0)
-	drawHeaderSeparatorLine(pdf, miniHeaderSeparatorBottomMargin)
-}
-
-// registerLogoImage registers logo bytes with the PDF
-func registerLogoImage(pdf *gofpdf.Fpdf, logoBytes []byte) (string, *gofpdf.ImageInfoType) {
-	_, format, err := image.Decode(bytes.NewReader(logoBytes))
-	if err != nil {
-		return "", nil
-	}
-	var imageType string
-	switch format {
-	case "jpeg", "jpg":
-		imageType = "JPEG"
-	case "png":
-		imageType = "PNG"
-	default:
-		return "", nil
-	}
-	opts := gofpdf.ImageOptions{ImageType: imageType}
-	info := pdf.RegisterImageOptionsReader("orglogo", opts, bytes.NewReader(logoBytes))
-	if info == nil {
-		return "", nil
-	}
-	return "orglogo", info
-}
-
-// drawHeaderSeparatorLine draws a horizontal line below the title/mini header and adds bottomMargin space below it.
-func drawHeaderSeparatorLine(pdf *gofpdf.Fpdf, bottomMargin float64) {
-	pdf.SetDrawColor(defaultBorderColor[0], defaultBorderColor[1], defaultBorderColor[2])
-	pdf.SetLineWidth(headerSeparatorLineWidth)
-	y := pdf.GetY()
-	pdf.Line(marginLR, y, mmPerPage-marginLR, y)
-	pdf.SetY(y)
-	pdf.Ln(bottomMargin)
-	pdf.SetLineWidth(0.25)
-	pdf.SetDrawColor(0, 0, 0)
-}
-
-// drawTitleSection draws the title block
-func drawTitleSection(pdf *gofpdf.Fpdf, walletAccount string) {
-	walletAddr := walletAccount
-	if strings.HasPrefix(walletAddr, "stellar:") {
-		walletAddr = walletAddr[8:]
-	}
-
-	pdf.SetFont("Inter", "", bodyFontSize)
-	pdf.SetTextColor(noteColor[0], noteColor[1], noteColor[2])
-	pdf.CellFormat(0, titleSectionLine1Height, "REPORT", "", 1, "L", false, 0, "")
-
-	pdf.SetFont("Inter", "B", titleFontSize)
-	pdf.SetTextColor(highlightColor[0], highlightColor[1], highlightColor[2])
-	pdf.CellFormat(0, titleSectionLine2Height, "Wallet Statement", "", 1, "L", false, 0, "")
-
-	// Wallet Address label and value using absolute positioning
-	yWalletLine := pdf.GetY()
-	pdf.SetFont("Inter", "", organizationNameFontSize)
-	pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-	labelText := "Wallet Address: "
-	labelWidth := pdf.GetStringWidth(labelText)
-	xLabelStart := pdf.GetX()
-	// Draw label at absolute position
-	pdf.SetXY(xLabelStart, yWalletLine)
-	pdf.CellFormat(labelWidth, titleSectionLine3Height, labelText, "", 0, "L", false, 0, "")
-	xValueStart := xLabelStart + labelWidth + walletAddressLabelGap
-	// Draw value at absolute position with different font
-	pdf.SetFont("GoogleSansCode", "U", organizationNameFontSize)
-	pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-	pdf.SetXY(xValueStart, yWalletLine)
-	walletURL := fmt.Sprintf("https://stellar.expert/explorer/testnet/account/%s", walletAddr)
-	walletAddrWidth := pdf.GetStringWidth(walletAddr)
-	pdf.CellFormat(walletAddrWidth, titleSectionLine3Height, walletAddr, "", 0, "L", false, 0, walletURL)
-	// Add link over the entire wallet address area
-	pdf.LinkString(xValueStart, yWalletLine, walletAddrWidth, titleSectionLine3Height, walletURL)
-	pdf.SetFont("Inter", "", organizationNameFontSize)
-	pdf.SetTextColor(defaultCellColor[0], defaultCellColor[1], defaultCellColor[2])
-	pdf.Ln(titleSectionLine3Height)
-
-	pdf.Ln(titleSectionBottomMargin)
-	pdf.SetTextColor(0, 0, 0)
-}
-
-// breakHeaderWords inserts newlines between words
-func breakHeaderWords(s string) string {
-	return strings.ReplaceAll(s, " ", "\n")
-}
-
-// getMockStatementResult returns a full statement with multiple currencies and mock transactions (for PDF testing).
-func getMockStatementResult() *services.StatementResult {
-	return &services.StatementResult{
-		Summary: services.StatementSummary{
-			Account: "stellar:GDNRRK5EXMZ4STV7UTO3CW4LSVNY5KYWTM3J7BM5SQNA7KE2RYX55IYV",
-			Assets: []services.StatementAssetSummary{
-				{
-					Code:             "XLM",
-					BeginningBalance: "4999.8900000",
-					TotalCredits:     "6000.0000000",
-					TotalDebits:      "1200.5000000",
-					EndingBalance:    "9799.3900000",
-					Transactions:     generateMockTransactions(8),
-				},
-				{
-					Code:             "USDC",
-					BeginningBalance: "10000.0000000",
-					TotalCredits:     "2500.5000000",
-					TotalDebits:      "3500.2500000",
-					EndingBalance:    "9000.2500000",
-					Transactions:     generateMockTransactions(12),
-				},
-				{
-					Code:             "EURT",
-					BeginningBalance: "500.0000000",
-					TotalCredits:     "150.7500000",
-					TotalDebits:      "200.0000000",
-					EndingBalance:    "450.7500000",
-					Transactions:     generateMockTransactions(5),
-				},
-			},
-		},
-	}
-}
-
-// generateMockTransactions returns count copies of a template transaction with slight variations (for testing).
-func generateMockTransactions(count int) []services.StatementTransaction {
-	baseTime, _ := time.Parse(time.RFC3339, "2026-01-29T15:00:10Z")
-	out := make([]services.StatementTransaction, 0, count)
-	for i := 0; i < count; i++ {
-		txType := "debit"
-		if i%2 == 1 {
-			txType = "credit"
-		}
-		createdAt := baseTime.Add(time.Duration(i) * time.Hour).UTC().Format(time.RFC3339)
-		out = append(out, services.StatementTransaction{
-			ID:                  "7cb4a68dc164ad69c6121086cf3aef0cec0d78634f60e1a1e23e4637b1f082e2",
-			CreatedAt:           createdAt,
-			UpdatedAt:           createdAt,
-			Type:                txType,
-			Amount:              "100.1000000",
-			CounterpartyAddress: "GAHSWJ2ANIFE3ZEWM4EN7WKLC2F4OCLS2O4QQQJSADYHOXZDA3EZNJ2M",
-			CounterpartyName:    "owner@bluecorp.local",
-			ExternalPaymentID:   "ext-payment-001",
-		})
-	}
-	return out
+	return statement.BuildPDF(result, fromDate, toDate, organizationName, organizationLogo)
 }
