+++ internal/pdf/fonts.go
@@ -1,0 +1,18 @@
+package pdf
+
+import _ "embed"
+
+//go:embed assets/fonts/Inter_24pt-Regular.ttf
+var interRegularFont []byte
+
+//go:embed assets/fonts/Inter_24pt-Bold.ttf
+var interBoldFont []byte
+
+//go:embed assets/fonts/Inter_24pt-Medium.ttf
+var interMediumFont []byte
+
+//go:embed assets/fonts/Inter_24pt-SemiBold.ttf
+var interSemiBoldFont []byte
+
+//go:embed assets/fonts/GoogleSansCode-Regular.ttf
+var googleSansCodeRegularFont []byte
