+++ internal/pdf/statement/utils.go
@@ -1,0 +1,7 @@
+package statement
+
+import "strings"
+
+func breakHeaderWords(s string) string {
+	return strings.ReplaceAll(s, " ", "\n")
+}
